Extract basic source and test files
